pyproject.toml
[project]
name = "moomoo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "matplotlib>=3.10.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
